squash msal
