--- msal/account_provider.go
@@ -1,12 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-package msal
-
-//AccountProvider is an interface representing an account that is returned to users.
-//This can help with accessing the cache for tokens.
-type AccountProvider interface {
-	GetUsername() string
-	GetHomeAccountID() string
-	GetEnvironment() string
-}
