--- msal/acquire_token_device_code_parameters_test.go
@@ -1,25 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-package msal
-
-import (
-	"reflect"
-	"testing"
-
-	"github.com/AzureAD/microsoft-authentication-library-for-go/internal/msalbase"
-)
-
-func TestAugmentAuthenticationParametersDeviceCode(t *testing.T) {
-	testScopes := []string{"user.read"}
-	testAuthParams := msalbase.AuthParametersInternal{}
-	testTokenCommonParams := acquireTokenCommonParameters{testScopes}
-	testDeviceCodeParams := acquireTokenDeviceCodeParameters{
-		commonParameters: testTokenCommonParams,
-	}
-	testDeviceCodeParams.augmentAuthenticationParameters(&testAuthParams)
-	actualScopes := testAuthParams.Scopes
-	if !reflect.DeepEqual(testScopes, actualScopes) {
-		t.Errorf("Actual scopes %v differ from expected scopes %v", actualScopes, testScopes)
-	}
-}
