--- msal/authentication_result_provider.go
@@ -1,10 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-package msal
-
-// AuthenticationResultProvider contains the results of one token acquisition operation in PublicClientApplication
-// or ConfidentialClientApplication.
-type AuthenticationResultProvider interface {
-	GetAccessToken() string
-}
